+++ faddr/faddr.py
@@ -3,12 +3,10 @@
 import argparse
 import sys
 
-from rich import box
-from rich.table import Table
-
 from faddr import __version__, console, logger
 from faddr.database import Database
 from faddr.exceptions import FaddrSettingsFileFormatError
+from faddr.results import NetworkResult
 from faddr.settings import load_settings
 
 
@@ -38,6 +36,14 @@ def parse_cmd_args():
         "--description",
         action="store_true",
         help="Print description column",
+    )
+
+    parser.add_argument(
+        "-o",
+        "--output",
+        choices=("table", "json"),
+        default="table",
+        help="Output format, default is 'table'",
     )
     parser.add_argument(
         "-s",
@@ -69,51 +75,6 @@ def parse_input(input_data):
     return query
 
 
-def pretty_print_result(
-    result,
-    description=False,
-    color=False,
-    table=False,
-    query_number=1,
-):
-    """Print data with pretty formatting."""
-
-    header = result["headers"]["full"].copy()
-    if description is False:
-        header.remove("Description")
-    if query_number == 1:
-        header.remove("Query")
-
-    table = Table(
-        expand=table,
-        highlight=color,
-        header_style=None,
-        box=box.SQUARE if table else None,
-        safe_box=True,
-        padding=(0, 1, 0, 1) if table else (0, 2, 0, 0),
-    )
-
-    for column_name in header:
-        table.add_column(
-            column_name,
-            overflow=None,
-        )
-
-    for row in result["data"]:
-        # Fix bool and None values.
-        for key, value in row.items():
-            if value is None:
-                row[key] = "-"
-            elif isinstance(value, bool) and value:
-                row[key] = "[bold red]Yes"
-            elif isinstance(value, bool) and not value:
-                row[key] = "-"
-
-        table.add_row(*[str(row.get(cell_name, "-")) for cell_name in header])
-
-    console.print(table)
-
-
 def main():
     """Query database"""
 
@@ -125,8 +86,6 @@ def main():
         print(__version__)
         sys.exit(0)
 
-    query = parse_input(cmd_args.get("ip_address"))
-
     try:
         settings = load_settings(settings_file=cmd_args.get("settings_file"))
     except FaddrSettingsFileFormatError as err:
@@ -136,12 +95,10 @@ def main():
 
     database = Database(**settings.database.dict())
 
-    result = database.find_networks(query)
-
-    pretty_print_result(
-        result,
-        description=cmd_args.get("description", False),
-        color=cmd_args.get("color", False),
-        table=cmd_args.get("table", False),
-        query_number=len(query),
+    result = NetworkResult(database.find_networks(cmd_args.get("ip_address")))
+    result.print(
+        include_description=cmd_args.get("description"),
+        output=cmd_args.get("output"),
+        color=cmd_args.get("color"),
+        border=cmd_args.get("table"),
     )